+++ cupy/_statistics/histogram.py
