+++ cupy/_statistics/meanvar.py
