+++ cupy/_statistics/order.py
