--- cupy/statistics/__init__.py
@@ -1,8 +1,0 @@
-# Functions from the following NumPy document
-# https://docs.scipy.org/doc/numpy/reference/routines.statistics.html
-
-# "NOQA" to suppress flake8 warning
-from cupy.statistics import correlation  # NOQA
-from cupy.statistics import histogram  # NOQA
-from cupy.statistics import meanvar  # NOQA
-from cupy.statistics import order  # NOQA
